requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
google-generativeai==0.7.1
python-dotenv==1.0.1
PyMySQL==1.1.0
cryptography==41.0.7